refactor(TripInstruction): Move to 'instruction' package.
