+++ src/main/java/org/opentripplanner/middleware/triptracker/instruction/TransitLegInstruction.java
@@ -1,7 +1,6 @@
 package org.opentripplanner.middleware.triptracker.instruction;
 
 import org.opentripplanner.middleware.otp.response.Leg;
-import org.opentripplanner.middleware.triptracker.TripInstruction;
 
 public class TransitLegInstruction extends TripInstruction {
     protected Leg transitLeg;
