+++ src/main/java/org/opentripplanner/middleware/triptracker/instruction/TransitLegSummaryInstruction.java
@@ -1,7 +1,6 @@
 package org.opentripplanner.middleware.triptracker.instruction;
 
 import org.opentripplanner.middleware.otp.response.Leg;
-import org.opentripplanner.middleware.triptracker.TripInstruction;
 
 import java.util.Locale;
