+++ src/main/java/org/opentripplanner/middleware/triptracker/instruction/TripInstruction.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.middleware.triptracker;
+package org.opentripplanner.middleware.triptracker.instruction;
 
 import org.opentripplanner.middleware.otp.response.Place;
 import org.opentripplanner.middleware.otp.response.Step;
